+++ man/examples/exa_is_undergrad.R
@@ -1,19 +1,27 @@
-# select min required columns
+# Select min required columns
 term <- toy_term[, .(mcid, term)]
 course <- toy_course[, .(mcid, term_course)]
 degree <- toy_degree[, .(mcid, term_degree)]
 
-# labeling terms by group: undergrad & grad
+# Example of starting data frame
+term
+
+# Labeling terms by group: undergrad & grad
 term <- is_undergrad(term, midf_table = degree)
 course <- is_undergrad(course, midf_table = degree)
 degree <- is_undergrad(degree, midf_table = degree)
 
-# results
+# Example result
 term[order(-term_focus)]
-term[, .N, by = "term_focus"][order(-N)]
 
-course[order(-term_focus)]
-course[, .N, by = "term_focus"][order(-N)]
+# No change if added columns are redundant
+x <- is_undergrad(term, midf_table = degree)
+check_equiv_frames(term, x)
 
-degree[order(-term_focus)]
-degree[, .N, by = "term_focus"][order(-N)]
+# Filter to retain "undergraduate" rows only
+term <- term[term_focus == "undergrad"]
+course <- course[term_focus == "undergrad"]
+degree <- degree[term_focus == "undergrad"]
+
+# Example result
+term